django-auth/pyproject.toml
[tool.poetry]
name = "django-auth"
version = "0.1.0"
description = ""
authors = ["nasif raihan <nasif.raihan78@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
